requirements.txt
SQLAlchemy==2.0.41
python-dotenv==1.0.1
Werkzeug==3.1.0
pytest==8.3.5
PyJWT==2.8.0
click==8.1.7